Replace long() with int().
The long() and int() were made equivalent in Python 2.4 and long() and
the 'long' type was removed in Python 3.

+++ Products/DataCollector/plugins/zenoss/cmd/df.py
@@ -58,7 +58,7 @@ class df(CommandPlugin):
                 om.totalBlocks = 0
             else:
                 try:
-                    om.totalBlocks = long(tblocks)
+                    om.totalBlocks = int(tblocks)
                 except ValueError:
                     # Ignore this filesystem if what we thought was total
                     # blocks isn't a number.
